extracted dashboard-info-card to be a  single component
